refactor: 💡 refact code
