+++ packages/webpack-plugin-extract-css-assets/test/actual/relative-path/main.css
@@ -19,7 +19,7 @@ body {
       format('embedded-opentype'),
     url('cssassets/f7efc431f760b9e8a6125054c5f4a502.woff') format('woff'),
     url('cssassets/3a70f449494ab666efaf24b72d563431.ttf') format('truetype'),
-    url('//at.alicdn.com/t/font_377048_kjcjjyweu4sra4i.svg#iconfont')
+    url('cssassets/b1b72dfa538406f3897c40ab412b4183.xml')
       format('svg');
 }
